requirements.txt
jupyter
pytorch
torchvision
torchaudio
numpy
matplotlib
torchsummary
wandb